requirements.txt
tensorflow-gpu
numpy
editdistance
pandas
